graphs and fps

+++ src/log_t.h
@@ -1,5 +1,5 @@
 static constexpr size_t log_channels = 8;
-static constexpr size_t log_samples = (size_t) std_audio_sample_rate_hz;
+static constexpr size_t log_samples = std_audio_sample_rate_hz;
 
 typedef struct sample_s log_t[log_channels][log_samples];
